requirements.txt
# Core data processing - Python 3.9 compatible versions
pandas>=1.5.0,<2.1.0
numpy>=1.21.0,<1.25.0
geopandas>=0.12.0,<0.14.0
shapely>=1.8.0,<2.1.0

# Database and storage
sqlalchemy>=1.4.0,<2.1.0

# API and web requests
requests>=2.28.0
urllib3>=1.26.0,<2.0.0
tenacity>=8.0.0

# GTFS processing - reliable libraries
partridge>=1.1.0
pygtfs>=0.1.7

# Geospatial processing
pyproj>=3.4.0
fiona>=1.8.0

# Data validation and quality
cerberus>=1.3.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
tqdm>=4.64.0
loguru>=0.6.0

# Analysis tools
matplotlib>=3.5.0
seaborn>=0.11.0

scipy>=1.10.0
plotly>=5.15.0
folium>=0.14.0
contextily>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0

# Dashboard and visualization
streamlit>=1.28.0
streamlit-folium>=0.15.0
streamlit-elements>=0.1.0
streamlit-option-menu>=0.3.0
streamlit-aggrid>=0.3.0
streamlit-chat>=0.1.1

# Machine Learning (for future phases)
scikit-learn>=1.3.0
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
hdbscan>=0.8.33
umap-learn>=0.5.4
joblib>=1.3.0

# NLP and LangChain for Policy Assistant (FREE/Open-source only)
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
# sentence-transformers already included above for embeddings
# Using local models - no API costs